Rename tupledict.py to dicttuple.py
