+++ packages/syft/src/syft/protocol/protocol_version.json
@@ -345,7 +345,7 @@
       "DatasetPageView": {
         "1": {
           "version": 1,
-          "hash": "1c7042b2980c9c91b6cb1bbdc99417b1b7bc88b1030223107d15a8e0f60183d6",
+          "hash": "6741bd16dc6089d9deea37b1bd4e895152d1a0c163b8bdfe45280b9bfc4a1354",
           "action": "add"
         }
       },
